ai-server/requirements.txt
amqp==5.0.6
billiard==3.6.4.0
celery==5.1.1
certifi==2019.9.11
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==1.6.0
cycler==0.10.0
dm-tree==0.1.6
future==0.18.2
greenlet==1.1.0
gym==0.18.3
kiwisolver==1.3.1
kombu==5.1.0
matplotlib==3.4.2
numpy==1.17.4
opencv-python==4.5.2.54
pandas==1.2.5
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycparser==2.19
pygame==2.0.1
pyglet==1.5.15
pymunk==6.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
redis==3.5.3
scipy==1.7.0
six==1.13.0
SQLAlchemy==1.4.20
tensorflow-model-optimization==0.6.0
torch==1.9.0
tqdm==4.61.1
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
wincertstore==0.2
